cli: consolidate confirmDestructive into pkg/cliutil
Lift the four byte-identical (modulo signature) confirmDestructive
copies — cluster/internal/clusteropts/confirm.go,
settings/{vpn,backup,restore}/common.go — into a single source at
cli/pkg/cliutil/confirm.go::ConfirmDestructive. The clusteropts
copy becomes a one-line re-export so the existing cluster callers
don't have to rewire imports; settings callers switch to the cliutil
package directly.

Backup/restore call sites previously wrapped each prompt in
`if !assumeYes { confirmDestructive(...) }` because the local helper
was 3-arg; now that everything uses the 4-arg cliutil shape, the
wrapper goes away (the assumeYes short-circuit lives inside the
helper). Net deletion ~70 lines, no behavior change.

cli/pkg/cliutil/ is the new home for project-wide CLI helpers shared
across cmd/ctl/* subtrees; future extractions (e.g. auth-error
reformatting, dash/age formatters used outside cluster) should land
here so the import graph stays one-way: cmd/ctl/* -> cliutil.

Made-with: Cursor
